% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_dc.R
\name{classify_dc}
\alias{classify_dc}
\title{Decide if a given hospital decision criteria is to be classified as outlier
or non-outlier}
\usage{
classify_dc(
  dc,
  threshold_value,
  digits = 8,
  pathway = "reporting",
  alternative = "greater"
)
}
\arguments{
\item{dc}{A vector of computed decision criteria as returned
by \code{\link{compute_rate_dc}} and \code{\link{compute_oe_dc}}}

\item{threshold_value}{Threshold value for the classification of the decision
 criteria. For \code{pathway = "reporting"}, it is the ratio of utilities
 \eqn{\frac{\mu}{\mu + \eta},} with  \eqn{\mu} denoting the utility of a correct
 outlier classification and \eqn{\eta} as utility of a correct non-outlier
classification. For \code{pathway = "change"}, it is the number of patients
with event of interest above the number which is tolerated by the reference
area. For \code{pathway = "naive"}, it is the reference value of
the indicator}

\item{digits}{Number of digits the result is rounded to before classification}

\item{pathway}{The pathway used for calculating the expected loss (either
\code{"reporting"} for the pathway of reporting or \code{"change"} for
the pathway of change)}

\item{alternative}{direction of the alternative (a character vector with
elements \code{"greater"} or \code{"less"}). The default value is
\code{"less"}.}
}
\value{
A vector of binary classification decisions indicating whether a
result is an outlier (\code{TRUE}) or non-outlier (\code{FALSE})
}
\description{
Decide if a given hospital decision criteria is to be classified as outlier
or non-outlier
}
\examples{
classify_dc(
  dc = 0.048, threshold_value = 0.05
)
}
