% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_oe_dc.R
\name{compute_oe_dc}
\alias{compute_oe_dc}
\title{Compute the decision criteria (dc) for classifying hospital results based on
binary quality indicators with risk adjustment}
\usage{
compute_oe_dc(
  o,
  e,
  t,
  alternative = "greater",
  prior_a = 0.5,
  prior_b = 0,
  value_if_e0 = NA,
  warn_if_e0 = TRUE
)
}
\arguments{
\item{o}{the observed number of events (a vector of non-negative integers)}

\item{e}{the expected number of events (a vector of non-negative numbers)}

\item{t}{the reference value}

\item{alternative}{direction of the alternative (a character vector with
elements \code{"greater"} or \code{"less"}). The default value is
\code{"less"}.}

\item{prior_a}{value for parameter a in the Gamma-distributed prior}

\item{prior_b}{value for parameter b in the Gamma-distributed prior}

\item{value_if_e0}{The computed decision criteria for the case
\code{o = e = 0} (a numeric of length one or \code{NULL}).  The default
value is \code{NA_real_}. If equal to \code{NULL}, then the computed decision
 criteria is equal to 1}

\item{warn_if_e0}{logical of length one.  If \code{TRUE} (the default), then
print a warning if the combination \code{o = e = 0} occurs in the input.}
}
\value{
A vector of computed decision criteria
}
\description{
Compute the decision criteria (dc) for classifying hospital (or other
healthcare providers) results based on binary quality indicators with risk
adjustment.
For the computation of the decision criteria, the number of "interesting"
events O is assumed to follow a Poisson distribution
with parameter \eqn{\theta \cdot e}, with \eqn{e} as the number of expected
events. The ratio O/E is usually referred to as standardized
mortality/morbidity ratio (SMR). The prior distribution of \eqn{\theta} is
 a Gamma-distribution with parameters \code{a = priori_a} and
\code{b = priori_b}. For risk adjusted indicators and the pathway of
reporting, the utility function is based on a
generalized 0-1 utility, with \eqn{\mu} denoting the utility of a correct
outlier classification and \eqn{\eta} as utility of a correct non-outlier
classification. In this case, the decision criteria is equivalent to the
posterior probability \deqn{\mbox{Pr}(\theta \le R|o,e).} For risk
adjusted O/E-indicators, only the decision criteria for the pathway of
reporting is currently implemented.
}
\examples{
compute_oe_dc(
  o = 1, e = 0.5, t = 1.5,
  alternative = "less"
)
}
\references{
Hengelbrock J, Rauh J, Cederbaum J, Kähler M, Höhle M (2021),
Hospital profiling using Bayesian decision theory,
medRxiv 2021.06.23.21259367; doi: https://doi.org/10.1101/2021.06.23.21259367
}
