% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rate_dc.R
\name{compute_rate_dc}
\alias{compute_rate_dc}
\title{Compute the decision criteria (dc) for classifying hospital results based on
binary quality indicators without risk adjustment}
\usage{
compute_rate_dc(
  o,
  n,
  t,
  alternative = "greater",
  prior_a = 0.5,
  prior_b = 0.5,
  pathway = "reporting",
  value_if_n0 = NA,
  warn_if_n0 = TRUE
)
}
\arguments{
\item{o}{the observed number of events (a vector of non-negative integers)}

\item{n}{total number of cases (a vector of non-negative integers that
satisfies \code{o <= n})}

\item{t}{the reference value (a vector of numbers
between 0 and 1)}

\item{alternative}{direction of the alternative (a character vector with
elements \code{"greater"} or \code{"less"}). The default value is
\code{"less"}.}

\item{prior_a}{value for parameter a in the Beta-distributed prior}

\item{prior_b}{value for parameter b in the Beta-distributed prior}

\item{pathway}{The pathway for which the decision criteria is calculated
(either \code{"reporting"} or \code{"change"})}

\item{value_if_n0}{The computed decision criteria for the case
\code{o = n = 0} (a numeric of length one or \code{NULL}).  The default
value is \code{NA_real_}. If equal to \code{NULL}, then the computed decision
 criteria is equal to 1}

\item{warn_if_n0}{logical of length one.  If \code{TRUE} (the default), then
print a warning if the combination \code{o = n = 0} occurs in the input.}
}
\value{
A vector of computed decision criteria
}
\description{
Compute the decision criteria (dc) for classifying hospital (or other
healthcare providers) results based on binary quality indicators without risk
adjustment.
For the computation of the decision criteria, the number of "interesting"
events O is assumed to follow a binomial distribution with success
probability \eqn{\theta}, \eqn{n} number of trials and with a
Beta-distributed prior with parameters \code{a = priori_a} and
\code{b = priori_b}. This type of indicator is called a "rate indicator".
In case of \code{pathway = "reporting"}, the utility function is based on a
generalized 0-1 utility, with \eqn{\mu} denoting the utility of a correct
outlier classification and \eqn{\eta} as utility of a correct non-outlier
classification. In this case, the decision criteria is equivalent to the
posterior probability \deqn{\mbox{Pr}(\theta \le R|o,n).}
In case of \code{pathway = "change"}, the utility of a correct outlier
classification depends upon the parameter \eqn{\theta} and the number of
patients \eqn{n}, with the decision criteria calculated as
\deqn{\int_{0}^{1} (\theta-R)_+ \cdot n \cdot f(\theta, o, n) d\theta.}
}
\examples{
compute_rate_dc(
  o = 1, n = 2, t = 0.95,
  alternative = "less"
)
}
\references{
Hengelbrock J, Rauh J, Cederbaum J, Kähler M, Höhle M (2021),
Hospital profiling using Bayesian decision theory,
medRxiv 2021.06.23.21259367; doi: https://doi.org/10.1101/2021.06.23.21259367
}
