% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_example.R
\name{run_example}
\alias{run_example}
\title{Run interactive examples for a funnelplot and the analysis of sensitivity and
specificity}
\usage{
run_example(example)
}
\arguments{
\item{example}{Either \code{"funnelplot"} for an interactive version of the
funnelplot or \code{"sensitivity_specificity"} for the interactive analysis
of sensitivity and specificity.}
}
\description{
Interactive versions of the \href{https://iqtig.shinyapps.io/funnel_plot/}{funnelplot}
and the \href{https://iqtig.shinyapps.io/sensitivity_specificity}{analysis of sensitivity and specificity}
are available online.
}
\examples{
if (interactive()) {
  run_example("funnelplot")
}
}
\references{
Spiegelhalter, D J  (2005),
Funnel plots for comparing institutional performance,
Statistics in Medicine 24(8):1185--1202
}
